backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
redis
chromadb
pydantic
twilio
